gas: castVote reduction
by making `updatePeriodEndBlock` smaller

+++ packages/nouns-contracts/contracts/governance/NounsDAOInterfaces.sol
@@ -703,7 +703,7 @@ contract NounsDAOStorageV3 {
         /// @notice The block at which this proposal was created
         uint256 creationBlock;
         address[] signers;
-        uint256 updatePeriodEndBlock;
+        uint64 updatePeriodEndBlock;
         uint64 objectionPeriodEndBlock;
         /// @notice When true, a proposal would be executed on timelockV1 instead of the current timelock
         bool executeOnTimelockV1;
